Add initial FASM generator support
At the moment, we need to still come up with a good enhancement of the
schema to support FASM features annotations to be added to the Device
Resources for a more data-driven approach.

With the current initial implementation, FASM features a quite fixed,
and, even though they can still be made more data-driven, some
specialization may be required to handle different devices families,
hence the adoption of a inheritance based generator, such that different
families can change the implementation or have more detailed functions
to add and deal with some possibly complex FASM annotations.

Signed-off-by: Alessandro Comodi <acomodi@antmicro.com>