Macro cluster generation
Add:
* Simple check if macro can be placed inside single site.
It only checks constraints, and bell counts. It doesn't check if routing exists from one bell to another one.
This may cause isses for some macros
* Finidng connecction graph. every cell is a node in this graph and
edge defines that cell's port is connected to other cell's port with the same net.
This is used in nextpnr to check if cells obtained from macro create given cluster as well as
to create routing when cluster is being placed
* BBA output for modified cluster definition

Fix:
* LDPE/LDCE FF_MODE from FF to LATCH, as these cell are latches

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>